refactor(graph): move JSON-LD schemas to their own namespace
