+++ renku/core/commands/schema/annotation.py
@@ -15,25 +15,24 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Represent an annotation for a workflow."""
+"""Annotation JSON-LD schema."""
 
-import copy
-from uuid import uuid4
+from marshmallow import EXCLUDE
+
+from renku.core.commands.schema.calamus import JsonLDSchema, dcterms, fields, oa
+from renku.core.models.provenance.annotation import Annotation
 
 
-class Annotation:
-    """Represents a custom annotation for a research object."""
+class AnnotationSchema(JsonLDSchema):
+    """Annotation schema."""
 
-    def __init__(self, *, id: str, body=None, source=None):
-        self.id = id
-        self.body = body
-        self.source = source
+    class Meta:
+        """Meta class."""
 
-    def copy(self):
-        """Return a copy of this annotation."""
-        return copy.copy(self)
+        rdf_type = oa.Annotation
+        model = Annotation
+        unknown = EXCLUDE
 
-    @staticmethod
-    def generate_id():
-        """Generate an id for an annotation."""
-        return f"/annotations/{uuid4().hex}"
+    id = fields.Id()
+    body = fields.RawJsonLD(oa.hasBody)
+    source = fields.Raw(dcterms.creator)
