+++ renku/core/models/dataset.py
@@ -27,15 +27,13 @@ from urllib.parse import quote, urlparse
 from uuid import uuid4
 
 import marshmallow
-from marshmallow import EXCLUDE
 
 from renku.core import errors
 from renku.core.metadata.database import Persistent
 from renku.core.metadata.immutable import Immutable, Slots
-from renku.core.models.calamus import DateTimeList, JsonLDSchema, Nested, Uri, fields, oa, prov, renku, schema
-from renku.core.models.entity import CollectionSchema, Entity, EntitySchema
-from renku.core.models.provenance.agent import Person, PersonSchema, SoftwareAgent
-from renku.core.models.provenance.annotation import Annotation, AnnotationSchema
+from renku.core.models.entity import Entity
+from renku.core.models.provenance.agent import Person, SoftwareAgent
+from renku.core.models.provenance.annotation import Annotation
 from renku.core.utils.datetime8601 import fix_datetime, local_now, parse_date
 from renku.core.utils.git import get_entity_from_revision
 from renku.core.utils.metadata import is_external_file
@@ -139,10 +137,6 @@ class DatasetTag(Persistent):
         identifier = Path(dataset_id).name
         name = quote(f"{name}@{identifier}", safe="")
         return f"/dataset-tags/{name}"
-
-    def to_jsonld(self):
-        """Create JSON-LD."""
-        return DatasetTagSchema().dump(self)
 
 
 class Language(Immutable):
@@ -296,10 +290,6 @@ class DatasetFile(Slots):
         """Return true if dataset is removed and should not be accessed."""
         return self.date_removed is not None
 
-    def to_jsonld(self):
-        """Create JSON-LD."""
-        return DatasetFileSchema(flattened=True).dump(self)
-
 
 class Dataset(Persistent):
     """Represent a dataset."""
@@ -366,15 +356,6 @@ class Dataset(Persistent):
         self.title: str = title
         self.version: str = version
         self.annotations: List[Annotation] = annotations or []
-
-    @classmethod
-    def from_jsonld(cls, data, schema_class=None):
-        """Create an instance from JSON-LD data."""
-        assert isinstance(data, (dict, list)), f"Invalid data type: {data}"
-
-        schema_class = schema_class or DatasetSchema
-        self = schema_class(flattened=True).load(data)
-        return self
 
     @staticmethod
     def generate_id(identifier: str) -> str:
@@ -588,139 +569,6 @@ class Dataset(Persistent):
     def clear_files(self):
         """Remove all files."""
         self.dataset_files = []
-
-    def to_jsonld(self):
-        """Create JSON-LD."""
-        return DatasetSchema(flattened=True).dump(self)
-
-
-class UrlSchema(JsonLDSchema):
-    """Url schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = schema.URL
-        model = Url
-        unknown = EXCLUDE
-
-    id = fields.Id(missing=None)
-    url = Uri(schema.url, missing=None)
-
-
-class DatasetTagSchema(JsonLDSchema):
-    """DatasetTag schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = schema.PublicationEvent
-        model = DatasetTag
-        unknown = EXCLUDE
-
-    dataset_id = Nested(schema.about, UrlSchema, missing=None)
-    date_created = fields.DateTime(schema.startDate, missing=None, format="iso", extra_formats=("%Y-%m-%d",))
-    description = fields.String(schema.description, missing=None)
-    id = fields.Id()
-    name = fields.String(schema.name)
-
-
-class LanguageSchema(JsonLDSchema):
-    """Language schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = schema.Language
-        model = Language
-        unknown = EXCLUDE
-
-    alternate_name = fields.String(schema.alternateName)
-    name = fields.String(schema.name)
-
-
-class ImageObjectSchema(JsonLDSchema):
-    """ImageObject schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = schema.ImageObject
-        model = ImageObject
-        unknown = EXCLUDE
-
-    content_url = fields.String(schema.contentUrl)
-    id = fields.Id(missing=None)
-    position = fields.Integer(schema.position)
-
-
-class RemoteEntitySchema(JsonLDSchema):
-    """RemoteEntity schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = [prov.Entity]
-        model = RemoteEntity
-        unknown = EXCLUDE
-
-    checksum = fields.String(renku.checksum)
-    id = fields.Id()
-    path = fields.String(prov.atLocation)
-    url = fields.String(schema.url)
-
-
-class DatasetFileSchema(JsonLDSchema):
-    """DatasetFile schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = [prov.Entity, schema.DigitalDocument]
-        model = DatasetFile
-        unknown = EXCLUDE
-
-    based_on = Nested(schema.isBasedOn, RemoteEntitySchema, missing=None)
-    date_added = DateTimeList(schema.dateCreated, format="iso", extra_formats=("%Y-%m-%d",))
-    date_removed = fields.DateTime(prov.invalidatedAtTime, missing=None, format="iso")
-    entity = Nested(prov.entity, [EntitySchema, CollectionSchema])
-    id = fields.Id()
-    is_external = fields.Boolean(renku.external, missing=False)
-    source = fields.String(renku.source, missing=None)
-
-
-class DatasetSchema(JsonLDSchema):
-    """Dataset schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = [prov.Entity, schema.Dataset]
-        model = Dataset
-        unknown = EXCLUDE
-
-    annotations = Nested(oa.hasTarget, AnnotationSchema, reverse=True, many=True)
-    creators = Nested(schema.creator, PersonSchema, many=True)
-    date_created = fields.DateTime(schema.dateCreated, missing=None, format="iso", extra_formats=("%Y-%m-%d",))
-    date_removed = fields.DateTime(prov.invalidatedAtTime, missing=None, format="iso")
-    date_published = fields.DateTime(
-        schema.datePublished, missing=None, format="%Y-%m-%d", extra_formats=("iso", "%Y-%m-%dT%H:%M:%S")
-    )
-    derived_from = Nested(prov.wasDerivedFrom, UrlSchema, missing=None)
-    description = fields.String(schema.description, missing=None)
-    dataset_files = Nested(schema.hasPart, DatasetFileSchema, many=True)
-    id = fields.Id(missing=None)
-    identifier = fields.String(schema.identifier)
-    images = fields.Nested(schema.image, ImageObjectSchema, missing=None, many=True)
-    in_language = Nested(schema.inLanguage, LanguageSchema, missing=None)
-    keywords = fields.List(schema.keywords, fields.String(), missing=None)
-    license = Uri(schema.license, missing=None)
-    name = fields.String(renku.slug)
-    initial_identifier = fields.String(renku.originalIdentifier)
-    project_id = fields.IRI(renku.hasDataset, reverse=True)
-    same_as = Nested(schema.sameAs, UrlSchema, missing=None)
-    title = fields.String(schema.name)
-    version = fields.String(schema.version, missing=None)
 
 
 class DatasetCreatorsJson(marshmallow.Schema):
