+++ renku/core/models/provenance/agent.py
@@ -21,11 +21,7 @@ import re
 import uuid
 from urllib.parse import quote
 
-from calamus.schema import JsonLDSchema
-from marshmallow import EXCLUDE
-
 from renku.core.metadata.immutable import Slots
-from renku.core.models.calamus import StringList, fields, prov, schema
 from renku.version import __version__, version_url
 
 
@@ -121,14 +117,6 @@ class Person(Agent):
         """Create and instance from a dictionary."""
         return cls(**data)
 
-    @classmethod
-    def from_jsonld(cls, data):
-        """Create an instance from JSON-LD data."""
-        if not isinstance(data, dict):
-            raise ValueError(data)
-
-        return PersonSchema().load(data)
-
     @staticmethod
     def generate_id(email, full_identity):
         """Generate identifier for Person."""
@@ -173,34 +161,3 @@ class Person(Agent):
     def full_identity(self):
         """Return name, email, and affiliation."""
         return self.get_full_identity(self.email, self.affiliation, self.name)
-
-
-class PersonSchema(JsonLDSchema):
-    """Person schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = [prov.Person, schema.Person]
-        model = Person
-        unknown = EXCLUDE
-
-    affiliation = StringList(schema.affiliation, missing=None)
-    alternate_name = StringList(schema.alternateName, missing=None)
-    email = fields.String(schema.email, missing=None)
-    id = fields.Id()
-    name = StringList(schema.name, missing=None)
-
-
-class SoftwareAgentSchema(JsonLDSchema):
-    """SoftwareAgent schema."""
-
-    class Meta:
-        """Meta class."""
-
-        rdf_type = [prov.SoftwareAgent]
-        model = SoftwareAgent
-        unknown = EXCLUDE
-
-    id = fields.Id()
-    name = StringList(schema.name, missing=None)
