+++ tests/core/models/test_shacl_schema.py
@@ -21,6 +21,7 @@ import pyld
 import pytest
 
 from renku.cli import cli
+from renku.core.commands.schema.dataset import dump_dataset_as_jsonld
 from renku.core.compat import Path
 from renku.core.utils.shacl import validate_graph
 from tests.utils import load_dataset
@@ -49,7 +50,7 @@ def test_dataset_shacl(tmpdir, runner, project, client):
     runner.invoke(cli, ["dataset", "tag", "dataset", "1.0"], catch_exceptions=False)
 
     dataset = load_dataset(client, "dataset")
-    g = dataset.to_jsonld()
+    g = dump_dataset_as_jsonld(dataset)
     rdf = pyld.jsonld.to_rdf(g, options={"format": "application/n-quads", "produceGeneralizedRdf": True})
 
     r, _, t = validate_graph(rdf, shacl_path=str(force_dataset_path))
@@ -67,7 +68,7 @@ def test_dataset_shacl(tmpdir, runner, project, client):
 
 def test_project_shacl(project, client, client_database_injection_manager):
     """Test project metadata structure."""
-    from renku.core.models.project import ProjectSchema
+    from renku.core.commands.schema.project import ProjectSchema
     from renku.core.models.provenance.agent import Person
 
     path = Path(__file__).parent.parent.parent / "data" / "force_project_shacl.json"
